Control.Exception.Annotation.Utils

Description

Provides utilities for working with GHC's native Exception annotations.

Since: 0.1

Global mechanisms

setUncaughtExceptionDisplay :: (String -> IO ()) -> IO ()

setUncaughtExceptionHandler with displayInnerMatchHandler i.e. calls displayException on any uncaught exceptions, passing the result to the param handler. ExitSuccess is ignored.

Since: 0.1

setUncaughtExceptionDisplayInnerMatch :: [ExceptionProxy] -> (String -> IO ()) -> IO ()

Like setUncaughtExceptionDisplay, except we attempt to match the given exception types.

Since: 0.1

Utils

data ExceptionProxy

Proxy for exception types. Used for matching multiple exception types.

Since: 0.1

Constructors

Exception e => MkExceptionProxy (Proxy e)

displayInner :: Exception e => e -> String

Calls displayException on SomeExceptions _inner_ exception i.e. given SomeException e calls displayException e. This means we will not use SomeException's default annotation printing.

Since: 0.1

displayInnerMatch :: Exception e => [ExceptionProxy] -> e -> String

displayInnerMatch proxies e attempts to convert e to one of the given exception types. If successful, we call displayException on the result. Otherwise calls displayException on the original parameter.

This can be useful when we do no want to use SomeExceptions displayException when e.g. we do not want to print annotations by default. For instance, we can use the following to ensure we do not print callstacks for specific Ex1 and Ex2:

data Ex1 = ... deriving anyclass Exception
data Ex2 = ... deriving anyclass Exception

let  displayEx' :: SomeException -> String
     displayEx' =
       displayInnerMatch
         [ MkExceptionProxy (Proxy @Ex1),
           MkExceptionProxy (Proxy @Ex2)
         ]

-- in main
setUncaughtExceptionHandler displayEx'

Since: 0.1

displayInnerMatchHandler

Arguments

:: Applicative f

=> [ExceptionProxy]
Proxies to attempt to match in displayInnerMatch.

-> (String -> f ())
The handler.

-> SomeException
The exception.

-> f ()
The result of the handler, or pure (), if we encountered ExitSuccess.

Calls displayInnerMatch on the exception, passing the result to the param handler. ExitSuccess is ignored.

Since: 0.1